refactor(infra): Move internal modules to infra namespace
