+++ auth/src/features/api/v1/mod.rs
@@ -1,6 +1,6 @@
 use axum::Router;
 
-use crate::{internal::config::Config, state::AppState};
+use crate::{infra::config::Config, state::AppState};
 
 pub mod auth;
 pub mod common;
