+++ auth/src/features/api/v1/users/info.rs
@@ -7,7 +7,7 @@ use crate::{
         user_info::{self, Gender},
         user_settings, users,
     },
-    internal::error::{AppError, AppErrorKind},
+    infra::error::{AppError, AppErrorKind},
     routers::{
         extractor::{LoginAccess, OperatorAccess},
         response::app_error_to_response,
