+++ auth/src/infra/config/common.rs
