+++ auth/src/infra/config/structure/mod.rs
