+++ auth/src/infra/error.rs
