+++ auth/src/infra/logger/mod.rs
