+++ auth/src/infra/mod.rs
@@ -1,1 +1,7 @@
+pub mod config;
 pub mod database;
+pub mod error;
+pub mod logger;
+pub mod mailer;
+pub mod session;
+pub mod utils;
