+++ auth/src/infra/utils/rand.rs
