--- auth/src/internal/mod.rs
@@ -1,17 +1,0 @@
-// Configuration module
-pub mod config;
-
-// Logging module
-pub mod log;
-
-// Error module
-pub mod error;
-
-// Session module
-pub mod session;
-
-// Utility module
-pub mod utils;
-
-// Mail module
-pub mod mail;
